Updated end points for user auth

+++ src/controller/group.controller.js
@@ -24,13 +24,16 @@ const getDevicesWithId = (req, res) => {
 
 const addNewGroup = (req, res) => {
     const schema = Joi.object({
+        "user_token": Joi.string().required(),
         "name": Joi.string().required()
     })
 
     const { error } = schema.validate(req.body);
     if(error) return res.status(400).send(error.details[0].message);
 
-    group.addNewGroup(req.body.name, (data, err) => {
+    let user_id = res.locals.user_id
+
+    group.addNewGroup(req.body.name, user_id, (data, err) => {
         if(err) res.sendStatus(500)
         res.status(201).send({"insertId": data.insertId});
     })
@@ -38,6 +41,7 @@ const addNewGroup = (req, res) => {
 
 const moveDeviceGroup = (req, res) => {
     const schema = Joi.object({
+        "user_token": Joi.string().required(),
         "mac_address": Joi.string().required()
     })
 
@@ -56,10 +60,24 @@ const moveDeviceGroup = (req, res) => {
 
         if(data[0].group_id == id) return res.sendStatus(304)
 
-        group.updateDeviceGroup(req.body.mac_address, id, (data, err) => {
+        //check group exists
+        group.getGroup(id, (data, err) =>{
+            if(err === 404) return res.sendStatus(404);
             if(err) return res.status(500).send(err);
-            return res.sendStatus(204)
+
+            let user_id = res.locals.user_id;
+            // check if authorised
+            if(data.connected_user != user_id){
+                return res.status(401).send("User unauthorised for this group");
+            }
+
+            group.updateDeviceGroup(req.body.mac_address, id, (data, err) => {
+                if(err) return res.status(500).send(err);
+                return res.sendStatus(204)
+            })
         })
+
+
     })
 
 }
@@ -71,6 +89,12 @@ const deleteGroup = (req, res) => {
     group.getGroup(id, (data, err) => {
         if(err === 404) return res.sendStatus(404)
         if(err) return res.sendStatus(500)
+
+        let user_id = res.locals.user_id;
+        // check if authorised
+        if(data.connected_user != user_id){
+            return res.status(401).send("User unauthorised for this group");
+        }
 
         group.deleteGroup(id, (data, err) => {
             if(err) return res.sendStatus(500);
